% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMethodToPeaks.R
\name{addMethodToPeaks}
\alias{addMethodToPeaks}
\title{add elution method name to rownames() to disambiguate results}
\usage{
addMethodToPeaks(sce, name = "method", sep = ":")
}
\arguments{
\item{sce}{a SingleCellExperiment with information about elution method.}

\item{name}{name of column or item indicating elution method ("method").}

\item{sep}{separator between method and peak name (":").}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a SingleCellExperiment with renamed rownames (or an error). 
}\if{html}{\out{</div>}}
}
\description{
This function will first search in metadata(sce)[\link{name}] for a value.
If none is found, the function will search for unique(colData(sce)[\link{name}]).
If neither works, the function will return an error.
}
