% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoModel.R
\name{autoModel}
\alias{autoModel}
\title{take an SCE of metabolomics data with factors in colData, add a model.matrix}
\usage{
autoModel(sce, ...)
}
\arguments{
\item{sce}{a SingleCellExperiment of mass spec data}
}
\description{
need some standardization of how experimental factors are conveyed in here
}
