% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEE_NMF.R
\name{iSEE_NMF}
\alias{iSEE_NMF}
\title{iSEE_NMF}
\usage{
iSEE_NMF(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment with NMF and UMAP reducedDims.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       nothing; an event loop will be initiated and iSEE() run
}\if{html}{\out{</div>}}
}
\description{
Simple wrapper function to launch an iSEE web application for an NMF'ed SCE.
iSEE is a Bioconductor package that you must install to run this function.
}
\examples{
if (require("iSEE")) {
  data(C18)
  iSEE_NMF(C18)
} 

}
\seealso{
SingleCellExperiment::SingleCellExperiment

RcppML::nmf

uwot::umap

iSEE::iSEE
}
