% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzR.R
\name{MzR}
\alias{MzR}
\title{accessor for assays(x, 'MzR')}
\usage{
MzR(x)
}
\arguments{
\item{x}{the object}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   assays(sce, "MzR") 
}\if{html}{\out{</div>}}
}
\description{
accessor for assays(x, 'MzR')
}
