% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeMzR.R
\name{normalizeMzR}
\alias{normalizeMzR}
\title{normalize metabolomics data with 'Exp', 'TC', and perhaps 'CAL' columns}
\usage{
normalizeMzR(sce)
}
\arguments{
\item{sce}{a SingleCellExperiment with !is.null(assay(sce, "MzR"))}
}
\description{
assay(sce, "normMzR") <- log1p((Exp|CAL) / TC) for all features in MzR
}
