% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normMzR.R
\name{normMzR}
\alias{normMzR}
\title{accessor for assays(x, 'normMzR')}
\usage{
normMzR(x)
}
\arguments{
\item{x}{the object}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   assays(sce, "normMzR") 
}\if{html}{\out{</div>}}
}
\description{
accessor for assays(x, 'normMzR')
}
