% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoNMF.R
\name{autoNMF}
\alias{autoNMF}
\title{take an SCE of metabolomics data with 'MzR' in the assays, norm, and NMF it}
\usage{
autoNMF(sce, k, ...)
}
\arguments{
\item{sce}{a SingleCellExperiment of mass spec data}

\item{k}{target rank for NMF decomposition}

\item{...}{additional arguments passed to RcppML::nmf}
}
\description{
This function assumes that TC and CAL samples have been used to normalize,
but if they have not (if there is no normMzR in assayNames(sce)), it will
run normalizeMzR() and the latter will complain if no TCs/CALs are present
}
