% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseColnames.R
\name{parseColnames}
\alias{parseColnames}
\title{parse MNP metabio standardized column names}
\usage{
parseColnames(cn, sep1 = "_", sep2 = "-")
}
\arguments{
\item{cn}{column names}

\item{sep1}{separator 1 ("_")}

\item{sep2}{separator 2 ("-")}

\item{drop1}{drop the first column? (TRUE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ a data.frame (coercible to DataFrame for SCE objects)
}\if{html}{\out{</div>}}
}
\description{
e.g. 'Exp_D9-8W-LN_2_C18neg-MC00138-S028'
splitting on '_':
token 1 is condition
token 2 is experimental factors, further split by '-':
factor1:'D9'
factor2:'8W'
factor3:'LN'
token 3 is replicate number (2)
token 4 is identifiers, further split by '-':
method:'C18neg'
experiment:'MC00138'
sample:'S028'
}
\details{
names(colData) for a default SCE would then be

c("experiment","method","sample","replicate","factor1","factor2","factor3")
}
\examples{
 
cn <- c("Exp_D9-8W-LN_1_C18neg-MC00138-S027", 
        "Exp_D9-8W-LN_2_C30neg-MC00138-S028")
        "Exp_D9-8W-LN_3_C18pos-MC00138-S029")
        "Exp_D9-8W-LN_4_308pos-MC00138-S030")
parseColnames(cn) 

}
