% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUntargeted.R
\name{readUntargeted}
\alias{readUntargeted}
\title{read untargeted metabolomics data into R and churn out an SCE}
\usage{
readUntargeted(x, coldat = TRUE, out = c("SCE", "raw"))
}
\arguments{
\item{x}{the CSV file of the output}

\item{coldat}{extract column data from the column names? (FALSE)}

\item{out}{format of the output object (default is "raw"; alt "SCE")}
}
\description{
Not to be confused with readMzR
}
\details{
To a first approximation, the function just reads in and parses colnames
}
\examples{

if (FALSE) { 
 
  library(Matrix)
  library(RcppML)
  library(singlet)

  fl <- "MC00295_HILIC_A_All_PoolSize.csv"
  HILAC <- readUntargeted(fl)
  HILAC <- HILAC[-which(rownames(HILAC) == "group"), ] 
  HILAC <- Matrix(log1p(HILAC)) # unstandardized 
  nmf_fit <- singlet::ard_nmf(HILAC)

}

}
